SSFChannel: change `sampling_rate` type to `float`
We typically take a `float` as it allows us to pass something like
`50e9` as well as an `int`.

+++ channel.py
@@ -71,7 +71,7 @@ class AWGN(Channel):
 class SSFChannel(Channel):
     h = 1000  # m
 
-    def __init__(self, fibre_length: int, sampling_rate: int) -> None:
+    def __init__(self, fibre_length: int, sampling_rate: float) -> None:
         super().__init__()
 
         assert fibre_length > 0
